remove notifications from user api calls and place into its own file
