+++ src/utilities/data/notificationConstants.ts
@@ -1,16 +1,17 @@
 import { NotificationInterface, NotificationState } from '../types/NotificationInterface';
 
-export const notificationInitialState: NotificationInterface = {
-  open: false,
-  type: 'info',
+export const notificationInitialState: NotificationState = {
   message: '',
+  open: false,
   timeout: 5000,
+  type: 'info',
 };
 
-export const emptyNotification: NotificationState = {
-  user: '',
-  notification: '',
+export const emptyNotification: NotificationInterface = {
   message: '',
   read: false,
+  title: '',
+  type: 0,
+  user: '',
   _id: '',
 };
