+++ src/utilities/redux/hooks.ts
@@ -1,12 +1,12 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import { NotificationActions } from './slices/users/notificationSlice';
 import { AppDispatch, RootState } from './store';
-import { NotificationInterface } from '../types/NotificationInterface';
+import { NotificationState } from '../types/NotificationInterface';
 
 export const useNotification = () => {
   const dispatch = useDispatch();
 
-  const displayNotification = (notification: NotificationInterface) => {
+  const displayNotification = (notification: NotificationState) => {
     dispatch(NotificationActions.addNotification(notification));
   };
