+++ src/utilities/redux/slices/users/notificationSlice.ts
@@ -1,12 +1,12 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { notificationInitialState } from '../../../data/notificationConstants';
-import { NotificationInterface } from '../../../types/NotificationInterface';
+import { NotificationState } from '../../../types/NotificationInterface';
 
 export const NotificationSlice = createSlice({
   name: 'notification',
   initialState: notificationInitialState,
   reducers: {
-    addNotification: (_state, action: PayloadAction<NotificationInterface>) => ({
+    addNotification: (_state, action: PayloadAction<NotificationState>) => ({
       ...notificationInitialState,
       ...action.payload,
       open: true,
