+++ src/utilities/redux/slices/users/userSlice.ts
@@ -11,7 +11,6 @@ const initialState: UiSliceInterface = {
       firstName: '',
       lastName: '',
       linkedinUrl: '',
-      notifications: [],
       portfolioUrl: '',
       profilePicture: '',
       role: '',
