+++ src/utilities/types/NotificationInterface.ts
@@ -1,16 +1,17 @@
 import { AlertColor } from '@mui/material';
 
-export interface NotificationInterface {
-  open?: boolean;
-  type?: AlertColor;
+export interface NotificationState {
   message?: string;
+  open?: boolean;
   timeout?: number | null;
+  type?: AlertColor;
 }
 
-export interface NotificationState {
-  user: string;
-  notification: string;
+export interface NotificationInterface {
   message: string;
   read: boolean;
+  title: string;
+  type: number;
+  user: string;
   _id: string;
 }
